Use Tox and setuptools
This is a more conventional interface for Python projects than `make`.
It allows the package to be installed with `pip`, and things like
`python setup.py test` are conventionally expected (so we don't need to
make an exception to the CircleCI configuration).

Running the tests in Tox isolates the test environment from the python
environment on the workstation. It would also be easy to edit tox.ini
and run the tests with Python 3, if that were a goal.